# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DexonApi",
    "version": "v1"
  },
  "paths": {
    "/api/Cml/GetAll": {
      "get": {
        "tags": [
          "Cml"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cml/by-info-id/{infoId}": {
      "get": {
        "tags": [
          "Cml"
        ],
        "parameters": [
          {
            "name": "infoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cml": {
      "post": {
        "tags": [
          "Cml"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cml"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cml"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cml"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Cml"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cml"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cml"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cml"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cml/{id}": {
      "delete": {
        "tags": [
          "Cml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Info/GetAll": {
      "get": {
        "tags": [
          "Info"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Info/{id}": {
      "get": {
        "tags": [
          "Info"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Info"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Info": {
      "post": {
        "tags": [
          "Info"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Info"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TestPoint/GetAll": {
      "get": {
        "tags": [
          "TestPoint"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TestPoint/by-cml-id/{cmlId}": {
      "get": {
        "tags": [
          "TestPoint"
        ],
        "parameters": [
          {
            "name": "cmlId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TestPoint/{id}": {
      "get": {
        "tags": [
          "TestPoint"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TestPoint"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TestPoint": {
      "post": {
        "tags": [
          "TestPoint"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestPoint"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TestPoint"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TestPoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "TestPoint"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestPoint"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TestPoint"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TestPoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Thickness/GetAll": {
      "get": {
        "tags": [
          "Thickness"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Thickness/by-tp-id/{tpId}": {
      "get": {
        "tags": [
          "Thickness"
        ],
        "parameters": [
          {
            "name": "tpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Thickness/{id}": {
      "get": {
        "tags": [
          "Thickness"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Thickness"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Thickness": {
      "post": {
        "tags": [
          "Thickness"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Thickness"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Thickness"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Thickness"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Thickness"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Thickness"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Thickness"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Thickness"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Info": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lineNumber": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "to": {
            "type": "string",
            "nullable": true
          },
          "drawingNumber": {
            "type": "string",
            "nullable": true
          },
          "service": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "inserviceDate": {
            "type": "string",
            "format": "date-time"
          },
          "pipeSize": {
            "type": "number",
            "format": "float"
          },
          "originalThickness": {
            "type": "number",
            "format": "float"
          },
          "stress": {
            "type": "number",
            "format": "float"
          },
          "jointEfficiency": {
            "type": "number",
            "format": "float"
          },
          "ca": {
            "type": "number",
            "format": "float"
          },
          "designLife": {
            "type": "number",
            "format": "float"
          },
          "designPressure": {
            "type": "number",
            "format": "float"
          },
          "operatingPressure": {
            "type": "number",
            "format": "float"
          },
          "designTemperature": {
            "type": "number",
            "format": "float"
          },
          "operatingTemperature": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "Cml": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "infoId": {
            "type": "integer",
            "format": "int32"
          },
          "info": {
            "$ref": "#/components/schemas/Info"
          },
          "lineNumber": {
            "type": "string",
            "nullable": true
          },
          "cmlNumber": {
            "type": "integer",
            "format": "int32"
          },
          "cmlDescription": {
            "type": "string",
            "nullable": true
          },
          "actualOutsideDiameter": {
            "type": "number",
            "format": "float"
          },
          "designThickness": {
            "type": "number",
            "format": "float"
          },
          "structuralThickness": {
            "type": "number",
            "format": "float"
          },
          "requiredThickness": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "TestPoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cmlId": {
            "type": "integer",
            "format": "int32"
          },
          "cml": {
            "$ref": "#/components/schemas/Cml"
          },
          "cmlNumber": {
            "type": "integer",
            "format": "int32"
          },
          "tpNumber": {
            "type": "integer",
            "format": "int32"
          },
          "tpDescription": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Thickness": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tpId": {
            "type": "integer",
            "format": "int32"
          },
          "testPoint": {
            "$ref": "#/components/schemas/TestPoint"
          },
          "tpNumber": {
            "type": "integer",
            "format": "int32"
          },
          "inspectionDate": {
            "type": "string",
            "format": "date-time"
          },
          "actualThickness": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      }
    }
  }
}